+++ src/pages/annotate/SmartTools/SmartToolsPanel.tsx
@@ -3,7 +3,8 @@ import { useDraggable } from '@neodrag/react';
 import { CircleX, FlaskConical, Grip, Images, Magnet, ScanText, ScissorsLineDashed, Sparkles, X } from 'lucide-react';
 import { LoadedImage } from '@/model';
 import { Button } from '@/ui/Button';
-import { useVisualSearchAvailable, VisualSearch } from './VisualSearch';
+import { useVisualSearchAvailable } from '@/utils/useVisualSearch';
+import { VisualSearch } from './VisualSearch';
 import { AnnotationMode, Tool } from '../AnnotationMode';
 import { AutoSelect } from './AutoSelect';
 import { EdgeSnap } from './EdgeSnap';
