+++ src/pages/annotate/SmartTools/VisualSearch/index.ts
@@ -1,2 +1,1 @@
-export * from './useVisualSearchAvailable';
 export * from './VisualSearch';
